Consume `OverdueSearchViewEffect` in `OverdueSearchEffectHandler`
