+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add layout for overdue search screen
 - Handle user interactions for overdue search results
 - Bump Sentry to v6.1.2
+- Consume `OverdueSearchViewEffect` in `OverdueSearchEffectHandler`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300
